# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/auth/refresh":{"get":{"operationId":"AuthController_refreshTokens","parameters":[],"responses":{"201":{"description":"Successful"},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/auth/signup":{"post":{"operationId":"AuthController_onUserSignUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupDTO"}}}},"responses":{"200":{"description":"successful"},"201":{"description":"account successfully created"},"400":{"description":"bad request"},"401":{"description":"unauthorized user"}},"tags":["Auth"]}},"/auth/account/verification":{"post":{"operationId":"AuthController_verifyEmail","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/otpDTO"}}}},"responses":{"201":{"description":"email verification successful"}},"tags":["Auth"]}},"/auth/resend/verification/otp":{"post":{"operationId":"AuthController_resendUserOTP","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/resendOTP"}}}},"responses":{"200":{"description":"successful"},"201":{"description":"resend verification otp"},"400":{"description":"Bad request"},"403":{"description":"forbidden response"}},"tags":["Auth"]}},"/auth/login":{"post":{"operationId":"AuthController_onUserLogin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}}},"responses":{"200":{"description":"successful"},"400":{"description":"bad request"},"409":{"description":"conflict responses"}},"tags":["Auth"]}},"/auth/email/forgot-password":{"post":{"operationId":"AuthController_sendEmailForgotPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailDTO"}}}},"responses":{"201":{"description":""},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"403":{"description":"forbidden response"}},"tags":["Auth"]}},"/auth/email/forgot-password/verify-otp":{"post":{"operationId":"AuthController_verifyPasswordResetOTP","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/otpDTO"}}}},"responses":{"201":{"description":"reset password otp successful"},"401":{"description":"Unauthorized"},"403":{"description":"forbidden response"},"404":{"description":"not found"}},"tags":["Auth"]}},"/auth/email/reset-password":{"post":{"operationId":"AuthController_setNewPassord","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}}},"responses":{"201":{"description":"password reset successful"},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"403":{"description":"forbidden response"}},"tags":["Auth"]}},"/user/profile":{"get":{"operationId":"UsersController_findOneUser","summary":"","description":"Get single user","parameters":[],"responses":{"200":{"description":"successful"},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"404":{"description":"Not found"}},"tags":["User"],"security":[{"bearer":[]}]}}},"info":{"title":"auth","description":"API Descriptions","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:8000"},{"url":"http://localhost:8080"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SignupDTO":{"type":"object","properties":{"first_name":{"type":"string","example":"Oyin"},"last_name":{"type":"string","example":"Vamp"},"password":{"type":"string","example":"Vamp@123"},"email":{"type":"string","example":"vamp@gmail.com"},"country_of_residence":{"type":"string","example":"Nigeria"},"occupation":{"type":"string","example":"Administration"}},"required":["first_name","last_name","password","email","country_of_residence","occupation"]},"otpDTO":{"type":"object","properties":{"email":{"type":"string","example":"vamp@gmail.com"},"otp":{"type":"string","example":"8790"}},"required":["email","otp"]},"resendOTP":{"type":"object","properties":{"email":{"type":"string","example":"vamp@gmail.com"},"otp":{"type":"string","example":"2345"}},"required":["email","otp"]},"LoginDTO":{"type":"object","properties":{"email":{"type":"string","description":"user unique email","example":"vamp@gmail.com"},"password":{"type":"string","example":"Vamp@123"}},"required":["email","password"]},"EmailDTO":{"type":"object","properties":{"email":{"type":"string","example":"vamp@gmail.com"}}}}}}